fix: Busy wait to avoid race condition
The CLI subcommand `format-device` encrypts any suitable file and
formats it as a BtrFS file system. For this it will, very briefly,
create an entry in `/dev/disk/by-uuid/`.

After exiting this subprogram the said entry will persist for a tiny
moment until Linux will tidy it up.

With version 2.11.0 `pydantic` became a bit faster here and there, that
is even claimed in the Changelog. Now the test for `format-device` was
fast enough to arrive at the creation of a symbolic link before the
system managed to remove the entry. This let to a raised exception,
since the link could not be created. As a consequence, the test failed.

To make matters worse, that crash only manifested itself on Github but
not locally. Therefore it took 21 commits in the final session to narrow
down the reason. That was extremely annoying!