fix: Fix behaviour of single files backup
Until now the backup of single files using the BtrFS backend was broken.
It would treat the destination to be a file if it missed a trailing
slash. As a consequence, only the last single file processed was backed
up, all others were overwritten.

After all the preparation in the preceding commits, this commit puts the
pieces together. It extends the tests so they would fail if single file
backup would not work correctly. Then it reactivates the rsync of single
files to FilesDest. Last but most importantly, it explicitly creates a
folder FilesDest, so `rsync` does the right thing.

On new backups this should ensure the correct behaviour.