test: Add test that reproduces a crash observed in the wild
This commit just adds a test for a bug that was very tricky to
reproduce. The test occured on existing backups after recent changes of
the single files handling, which caused these backups to break. Such
crashes are unacceptable! Old backups with valid configurations must
succeed!

This commit just adds the test and marks it as expected failure. Later
commits will refactor the test a bit and then add a fix.